refactor advisory policy sign off router
